requirements.txt
altgraph==0.17.4
customtkinter==5.2.2
darkdetect==0.8.0
numpy==2.1.3
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.0.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
setuptools==75.6.0
